Code cleanup and rename Readers folder to External
