+++ src/Application/Contracts/IConfigurationReader.cs
@@ -1,5 +1,3 @@
-using Merrsoft.MerrMail.Domain.Contracts;
-
 namespace Merrsoft.MerrMail.Application.Contracts;
 
 public interface IConfigurationReader
