+++ src/Infrastructure/External/EnvConfigurationReader.cs
@@ -1,9 +1,8 @@
 using Merrsoft.MerrMail.Application.Contracts;
 using Merrsoft.MerrMail.Domain.Contracts;
-using Merrsoft.MerrMail.Domain.Models;
 using Microsoft.Extensions.Logging;
 
-namespace Merrsoft.MerrMail.Infrastructure.Readers;
+namespace Merrsoft.MerrMail.Infrastructure.External;
 
 public class EnvConfigurationReader(
     ILogger<EnvConfigurationReader> logger,
